Start moving traits out of element (#38)
