+++ traits/src/lib.rs
@@ -1,5 +1,7 @@
 //! Trait definitions for the solvers library
 
+pub mod cell;
+pub mod element;
 pub mod fmm;
 pub mod general;
 pub mod tree;
